Add -recursive option
This unifies -name and -all to both not recurse by default, but provides the option to do so. Previously -name never recursed, while -all always recursed. This change is helpful for either finding a specific interface name in a multi-package project or limiting the depth of -all.

Breaking Change: Users depending on the previously recursive behavior of -all will need to add use -recursive to achieve the same.